LayoutTests/discovery/TC-7.9.4.5.6.1-expected.txt

Tests navigator.getNetworkServices() works in happy case as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices() first call to ok callback.
PASS srvs.servicesAvailable==0 is true
PASS srvs.servicesAvailable>=1 is true
PASS req.status==200 is true
PASS navigator.getNetworkServices() second call to ok callback.
PASS successfullyParsed is true

TEST COMPLETE
